food-waste-app/models/api/requirements.txt
fastapi
uvicorn
torch
pillow
transformers
peft
trl
datasets
bitsandbytes